# le-backend/docs/openapi_enhanced.yaml
openapi: 3.0.3
info:
  title: LC Workflow API - Enhanced
  description: |
    Enhanced LC Workflow API with comprehensive system stability improvements, 
    security features, monitoring capabilities, and data synchronization.
    
    ## Features
    
    - **Enhanced File Management**: Secure file upload with malware scanning and encryption
    - **Folder Organization**: Automatic document organization by type
    - **Health Monitoring**: Comprehensive system health checks and metrics
    - **Data Synchronization**: Real-time updates and cache management
    - **Security**: Advanced authentication, authorization, and audit logging
    - **Feature Flags**: Gradual rollout and A/B testing capabilities
    
    ## Authentication
    
    All endpoints require JWT authentication unless specified otherwise.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Error Handling
    
    The API uses standardized error responses with correlation IDs for tracking.
    All errors include suggested actions and detailed information.
    
    ## Rate Limiting
    
    - File Upload: 100 requests/hour per user
    - General API: 1000 requests/hour per user
    - Health Checks: 60 requests/minute (no user limit)
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /data-sync/cache/invalidate:
    post:
      tags:
        - Data Sync
      summary: Invalidate cache
      description: Manually invalidate cache for specific scope
      operationId: invalidateCache
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheInvalidationRequest'
      responses:
        '200':
          description: Cache invalidated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheInvalidationResponse'
  /data-sync/realtime/connect:
    post:
      tags:
        - Data Sync
      summary: Create real-time connection
      description: Create real-time update connection for SSE
      operationId: createRealtimeConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSEConnectionRequest'
      responses:
        '200':
          description: Connection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSEConnectionResponse'
  /data-sync/realtime/stream/{connection_id}:
    get:
      tags:
        - Data Sync
      summary: Real-time stream
      description: Server-Sent Events stream for real-time updates
      operationId: realtimeStream
      parameters:
        - name: connection_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
  /data-sync/verification/run:
    post:
      tags:
        - Data Sync
      summary: Run data verification
      description: Run data synchronization verification
      operationId: runDataVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        '200':
          description: Verification completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResponse'

  # Feature Flags Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CacheInvalidationRequest:
      type: object
      required:
        - scope
      properties:
        scope:
          type: string
          enum: [files, folders, applications, users]
        reason:
          type: string
          default: manual_refresh
        entity_id:
          type: string
          nullable: true
        related_ids:
          type: array
          items:
            type: string
    CacheInvalidationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        scope:
          type: string
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
    SSEConnectionRequest:
      type: object
      properties:
        subscribed_scopes:
          type: array
          items:
            type: string
          nullable: true
        application_filters:
          type: array
          items:
            type: string
          nullable: true
    SSEConnectionResponse:
      type: object
      properties:
        connection_id:
          type: string
        user_id:
          type: string
        subscribed_scopes:
          type: array
          items:
            type: string
        application_filters:
          type: array
          items:
            type: string
    VerificationRequest:
      type: object
      properties:
        scopes:
          type: array
          items:
            type: string
          nullable: true
        application_id:
          type: string
          nullable: true
        auto_fix:
          type: boolean
          default: false
    VerificationResponse:
      type: object
      properties:
        verification_results:
          type: object
        auto_fix_results:
          type: object
        total_issues:
          type: integer
        critical_issues:
          type: integer
        auto_fixable_issues:
          type: integer
        requested_by:
          type: string

    # Feature Flags Schemas
...